requirements.txt
python-dotenv
aiogram
requests